Install provided widgets into news scene
