+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Application/Components/News/Compositional/Component/CompositionalNewsComponentBuilder.swift
@@ -1,18 +1,24 @@
 public class CompositionalNewsComponentBuilder {
     
     private var sceneFactory: CompositionalNewsSceneFactory
+    private var widgets = [any NewsWidget]()
     
     public init() {
         sceneFactory = DefaultCompositionalNewsSceneFactory()
     }
     
-    public func with(_ sceneFactory: CompositionalNewsSceneFactory) -> Self {
+    public func with(_ sceneFactory: any CompositionalNewsSceneFactory) -> Self {
         self.sceneFactory = sceneFactory
         return self
     }
     
-    public func build() -> NewsComponentFactory {
-        return CompositionalNewsComponentFactory(sceneFactory: sceneFactory)
+    public func with(_ widget: any NewsWidget) -> Self {
+        self.widgets.append(widget)
+        return self
+    }
+    
+    public func build() -> any NewsComponentFactory {
+        return CompositionalNewsComponentFactory(sceneFactory: sceneFactory, widgets: widgets)
     }
     
 }
