+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Application/Components/News/Compositional/Component/CompositionalNewsComponentFactory.swift
@@ -2,10 +2,17 @@ import UIKit
 
 struct CompositionalNewsComponentFactory: NewsComponentFactory {
     
-    let sceneFactory: CompositionalNewsSceneFactory
+    let sceneFactory: any CompositionalNewsSceneFactory
+    let widgets: [any NewsWidget]
     
-    func makeNewsComponent(_ delegate: NewsComponentDelegate) -> UIViewController {
-        sceneFactory.makeCompositionalNewsScene()
+    func makeNewsComponent(_ delegate: any NewsComponentDelegate) -> UIViewController {
+        let compositionalNewsScene = sceneFactory.makeCompositionalNewsScene()
+        
+        for widget in widgets {
+            widget.register(in: compositionalNewsScene)
+        }
+        
+        return compositionalNewsScene
     }
     
 }
