+++ Packages/EurofurenceApplication/Sources/EurofurenceApplication/Application/Components/News/Legacy/Component/NewsComponentFactory.swift
@@ -3,7 +3,7 @@ import UIKit.UIViewController
 
 public protocol NewsComponentFactory {
 
-    func makeNewsComponent(_ delegate: NewsComponentDelegate) -> UIViewController
+    func makeNewsComponent(_ delegate: any NewsComponentDelegate) -> UIViewController
 
 }
 
@@ -11,7 +11,7 @@ public protocol NewsComponentDelegate {
 
     func newsModuleDidRequestShowingPrivateMessages()
     func newsModuleDidSelectAnnouncement(_ announcement: AnnouncementIdentifier)
-    func newsModuleDidSelectEvent(_ event: Event)
+    func newsModuleDidSelectEvent(_ event: any Event)
     func newsModuleDidRequestShowingAllAnnouncements()
 
 }
